add support for function to monitor activity of spot instances
We want a separate thread that can control watching the spot instances, and
keeping track of when they appear and go away. While not exact, if this runs
every 15 seconds we can get some sense of this! This function also
currently takes charge of disabling hyper threading via the hotplug
script. Later I am going to explore accomplishing this via a launch
template, so that we can have assurance that the nodes come up
ready to go. My biggest concern here is that the template will not
allow for detailed enough customization of different spot types
and the need for the single thread. We may need to run the logic
on all types, regardless.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>