update spot instance script
I am stopping here because I do not see that spot prices are
an improvement over the hpc instance types. E.g.,

 - hpc7g: 64 physical cores, $1.68 per hour
   - hpc6a: 96 physical cores, $2.88 per hour
and it seems like this maybe gives us 1-3 instance types
per filter group, assuming those are physical cores and
not vCPU. We should continue this if we are interested
in other instance sizes only.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>