performance-study: organizing 2 node tests and builds.
We are needing to estimate costs on GKE, and thus need to know
a quasi final application set and approximate times. This base of
work will build as many containers as we can and test them on 2 nodes
each with 4 GPU, meaning on GKE with miniclusters. When we know the
scale we will have, we will extend these into that scale, do
timing for the different applications and sizes, and prepare
for the larger experiment.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>