fixes: docker for gpu testing and test setup
This adds tweaks to the Google cloud Dockerfile intended for GKE,
specifically I noticed that we had mpich/lib installed, and I was
worried to have another MPI library pulled in (when we want our
custom build to be the only one). I rebuilt (still have 5 to go)
all the containers, and also added oras, which we need for the
experiment results saving. I am also adding the prototype test
run file, which is one set of instructions and common, shared
CRD that we can use to run the experiment (and test soon).

Signed-off-by: vsoch <vsoch@users.noreply.github.com>