performance: add container repository commits
It might be redundant, but I am looking in each
container to get the exact commit (and thus version).
It is important to note that some containers have
dependencies that I do not show here - we can look up
any that are needed. Finally, the source of truth for
build flags and additional configs or experiment files
should come from the Dockerfiles themselves.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>